+++ src/app/configs/config.ts
@@ -4,6 +4,7 @@ import {
   ACTIVATE_MENU,
   AUTO_IMPORT,
   CUSTOM_COMMANDS,
+  CUSTOM_TEMPLATES,
   EXCLUDE,
   INCLUDE,
   MenuInterface,
@@ -26,6 +27,7 @@ import {
  * @property {string[]} watch - The files to watch
  * @property {boolean} showPath - Whether to show the path or not
  * @property {object[]} customCommands - The custom commands
+ * @property {object[]} customTemplates - The custom templates
  * @property {object} activateItem - Whether to show the menu or not
  * @property {boolean} autoImport - The auto import setting
  * @property {boolean} skipFolderConfirmation - Whether to skip the folder confirmation or not
@@ -117,6 +119,20 @@ export class Config {
    */
   customCommands: object[];
   /**
+   * The custom templates.
+   * @type {object[]}
+   * @public
+   * @memberof Config
+   * @example
+   * const config = new Config(workspace.getConfiguration());
+   * console.log(config.customTemplates);
+   * console.log(config.customTemplates[0].name);
+   * console.log(config.customTemplates[0].description);
+   * console.log(config.customTemplates[0].type);
+   * console.log(config.customTemplates[0].template);
+   */
+  templates: object[];
+  /**
    * Whether to show the menu or not.
    * @type {MenuInterface}
    * @public
@@ -184,6 +200,10 @@ export class Config {
       'submenu.customCommands',
       CUSTOM_COMMANDS,
     );
+    this.templates = config.get<object[]>(
+      'submenu.templates',
+      CUSTOM_TEMPLATES,
+    );
     this.activateItem = config.get<MenuInterface>(
       'submenu.activateItem',
       ACTIVATE_MENU,
@@ -225,6 +245,10 @@ export class Config {
     this.customCommands = config.get<object[]>(
       'submenu.customCommands',
       CUSTOM_COMMANDS,
+    );
+    this.templates = config.get<object[]>(
+      'submenu.templates',
+      CUSTOM_TEMPLATES,
     );
     this.activateItem = config.get<MenuInterface>(
       'submenu.activateItem',